making create_page a Class, OcPage
We want to make every page (visible or invisible) have
the same style, alignment, etc. so, making create_page
to be a Class can just make it easier to do that. Now
plugins and pages not shown with tab can just use OcPage
to create itself, with the same style as those pages.

Signed-off-by: sunus <sunuslee@gmail.com>